#51 change column rate_card to bigint

+++ database/migrations/2017_09_13_125733_create_billboard_faces_table.php
@@ -18,7 +18,7 @@ class CreateBillboardFacesTable extends Migration
             $table->enum('facing', ['North', 'South', 'East', 'West', 'Other'])->nullable();
             $table->enum('reads', ['Left', 'Right', 'Across'])->nullable();
 
-            $table->decimal('rate_card', 10, 2)->default(0)->comment('the suggest price for this face');
+            $table->bigInteger('rate_card')->default(0)->comment('the suggest price for this face');
             $table->bigInteger('monthly_impressions')->default(0);
 
             $table->enum('type', ['Static', 'Digital'])->default('Static');
